Include action buttons inside Requests lists in a cleaner way, get rid of unnecessary JS when initializing DataTable
I initially used map() to create the action buttons and status badge... But this was unnecessary and overcomplicated the template, because it was done outside the "for p_request in request_list" loop, so I had to save the variables and then access them using JavaScript. It also made it more difficult to utilize the action_buttons.html as a separate template.