Create 2 Views, restrict view permissions, change display_type check
CRUD Views are now only accessible to logged-in users. RequestListView is now used as a parent for MyRequestListView and PendingRequestListView which override get_queryset to filter necessary objects. The DeleteView uses UserPassesTestMixin to check whether the person who tries to delete an object is the creator, and throws a modal with an error otherwise.